Factored out Sony's signature-generating algorithm and verified it works with PS3 saves
